+++ app/ante/ante.go
@@ -10,14 +10,9 @@ import (
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 	"github.com/cosmos/cosmos-sdk/types/tx/signing"
 	authante "github.com/cosmos/cosmos-sdk/x/auth/ante"
-	authsigning "github.com/cosmos/cosmos-sdk/x/auth/signing"
 	authtypes "github.com/cosmos/cosmos-sdk/x/auth/types"
 
-	channelkeeper "github.com/cosmos/ibc-go/v2/modules/core/04-channel/keeper"
-	ibcante "github.com/cosmos/ibc-go/v2/modules/core/ante"
-
 	"github.com/tharsis/ethermint/crypto/ethsecp256k1"
-	evmtypes "github.com/tharsis/ethermint/x/evm/types"
 )
 
 const (
@@ -28,15 +23,7 @@ const (
 // Ethereum or SDK transaction to an internal ante handler for performing
 // transaction-level processing (e.g. fee payment, signature verification) before
 // being passed onto it's respective handler.
-func NewAnteHandler(
-	ak evmtypes.AccountKeeper,
-	bankKeeper evmtypes.BankKeeper,
-	evmKeeper EVMKeeper,
-	feeGrantKeeper authante.FeegrantKeeper,
-	channelKeeper channelkeeper.Keeper,
-	feeMarketKeeper evmtypes.FeeMarketKeeper,
-	signModeHandler authsigning.SignModeHandler,
-) sdk.AnteHandler {
+func NewAnteHandler(options HandlerOptions) sdk.AnteHandler {
 	return func(
 		ctx sdk.Context, tx sdk.Tx, sim bool,
 	) (newCtx sdk.Context, err error) {
@@ -51,24 +38,11 @@ func NewAnteHandler(
 				switch typeURL := opts[0].GetTypeUrl(); typeURL {
 				case "/ethermint.evm.v1.ExtensionOptionsEthereumTx":
 					// handle as *evmtypes.MsgEthereumTx
-
-					anteHandler = sdk.ChainAnteDecorators(
-						NewEthSetUpContextDecorator(),                         // outermost AnteDecorator. SetUpContext must be called first
-						NewEthMempoolFeeDecorator(evmKeeper, feeMarketKeeper), // Check eth effective gas price against minimal-gas-prices
-						NewEthValidateBasicDecorator(evmKeeper),
-						NewEthSigVerificationDecorator(evmKeeper),
-						NewEthAccountVerificationDecorator(ak, bankKeeper, evmKeeper),
-						NewEthNonceVerificationDecorator(ak),
-						NewEthGasConsumeDecorator(evmKeeper),
-						NewCanTransferDecorator(evmKeeper, feeMarketKeeper),
-						NewEthIncrementSenderSequenceDecorator(ak), // innermost AnteDecorator.
-					)
-
+					anteHandler = newEthAnteHandler(options)
 				default:
 					return ctx, sdkerrors.Wrapf(
 						sdkerrors.ErrUnknownExtensionOptions,
-						"rejecting tx with unsupported extension option: %s",
-						typeURL,
+						"rejecting tx with unsupported extension option: %s", typeURL,
 					)
 				}
 
@@ -77,25 +51,9 @@ func NewAnteHandler(
 		}
 
 		// handle as totally normal Cosmos SDK tx
-
 		switch tx.(type) {
 		case sdk.Tx:
-			anteHandler = sdk.ChainAnteDecorators(
-				authante.NewSetUpContextDecorator(), // outermost AnteDecorator. SetUpContext must be called first
-				authante.NewRejectExtensionOptionsDecorator(),
-				authante.NewMempoolFeeDecorator(),
-				authante.NewValidateBasicDecorator(),
-				authante.NewTxTimeoutHeightDecorator(),
-				authante.NewValidateMemoDecorator(ak),
-				ibcante.NewAnteDecorator(channelKeeper),
-				authante.NewConsumeGasForTxSizeDecorator(ak),
-				authante.NewSetPubKeyDecorator(ak), // SetPubKeyDecorator must be called before all signature verification decorators
-				authante.NewValidateSigCountDecorator(ak),
-				authante.NewDeductFeeDecorator(ak, bankKeeper, feeGrantKeeper),
-				authante.NewSigGasConsumeDecorator(ak, DefaultSigVerificationGasConsumer),
-				authante.NewSigVerificationDecorator(ak, signModeHandler),
-				authante.NewIncrementSequenceDecorator(ak), // innermost AnteDecorator
-			)
+			anteHandler = newCosmosAnteHandler(options)
 		default:
 			return ctx, sdkerrors.Wrapf(sdkerrors.ErrUnknownRequest, "invalid transaction type: %T", tx)
 		}